(tck) deeply optimizing the JS code (tickets.js) (tck) securizing (early) the tck app (tck) correcting stuff to print properly "duplicata" on tickets (db)  adding a (Postgre)SQL function to be able to do sums of booleans

+++ branches/v2/lib/form/doctrine/TicketForm.class.php
@@ -42,9 +42,11 @@ class TicketForm extends BaseTicketForm
         ->andWhere('t.manifestation_id = ?', $this->object->manifestation_id)
         ->andWhere('t.transaction_id = ?', $this->object->transaction_id)
         ->andWhere('p.name = ?', $this->object->price_name)
+        ->andWhere('NOT t.printed')
         ->limit(1);
       $tickets = $q->execute();
-      $tickets[0]->delete();
+      if ( $tickets->count() > 0 )
+        $tickets[0]->delete();
       return array();
     }
     else
